set focus on the input field;  add confirmation via enter/return

+++ src/app/cluster/cluster-delete-confirmation/cluster-delete-confirmation.component.ts
@@ -1,31 +1,39 @@
 
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, DoCheck } from '@angular/core';
 import { Store } from "@ngrx/store";
 import * as fromRoot from "../../reducers/index";
 import { RouterModule, Router } from "@angular/router";
 import { ApiService } from "../../api/api.service";
 import { ClusterModel } from "../../api/model/ClusterModel";
 import { NotificationComponent } from "../../notification/notification.component";
+import { MdDialogRef } from '@angular/material';
 
 @Component({
   selector: 'kubermatic-cluster-delete-confirmation',
   templateUrl: './cluster-delete-confirmation.component.html',
   styleUrls: ['./cluster-delete-confirmation.component.scss']
 })
-export class ClusterDeleteConfirmationComponent implements OnInit {
-
+export class ClusterDeleteConfirmationComponent implements OnInit, DoCheck {
+  
   @Input() humanReadableName: string;
   @Input() clusterName: string;
   @Input() seedDcName: string;
-
+  
   public disableDeleteCluster: boolean = false;
   public clusterModel: ClusterModel;
   public cluster: any;
-
-  constructor(private router: Router, private api: ApiService, private store: Store<fromRoot.State>) {}
-
-  ngOnInit() {
-
+  
+  constructor(
+    private router: Router,
+    private api: ApiService, 
+    private store: Store<fromRoot.State>,
+    private dialogRef:MdDialogRef<ClusterDeleteConfirmationComponent>
+  ) {}
+  
+  ngOnInit() {}
+  
+  ngDoCheck(): void {
+    document.getElementById('name').focus();
   }
 
   onChange(event: any) {
@@ -37,14 +45,17 @@ export class ClusterDeleteConfirmationComponent implements OnInit {
   }
 
   deleteCluster(){
-    this.clusterModel = new ClusterModel(this.seedDcName, this.clusterName);
-    this.api.deleteCluster(this.clusterModel).subscribe(result => {
-      this.cluster = result;
-      NotificationComponent.success(this.store, "Success", `Cluster removed successfully`);
-
-      this.router.navigate(['/clusters']);
-    }, error => {
-      NotificationComponent.error(this.store, "Error", `${error.status} ${error.statusText}`);
-    })
+    if(this.disableDeleteCluster == true) {
+        this.dialogRef.close();
+        this.clusterModel = new ClusterModel(this.seedDcName, this.clusterName);
+        this.api.deleteCluster(this.clusterModel).subscribe(result => {
+          this.cluster = result;
+          NotificationComponent.success(this.store, "Success", `Cluster removed successfully`);
+    
+          this.router.navigate(['/clusters']);
+        }, error => {
+          NotificationComponent.error(this.store, "Error", `${error.status} ${error.statusText}`);
+        });
+    }
   }
 }
